+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Implemented the "$gt" and "$gte" query operations
+- Add parentheses for complex logical operations
 
 ### Fixed